requirements.txt

telebot==0.0.5
python-telegram-bot==20.7
python-dotenv==1.0.0


huggingface-hub==0.23.0
transformers==4.40.0
torch==2.3.0
sentence-transformers==2.7.0
accelerate==0.30.0
bitsandbytes==0.43.0


langchain==0.1.20
langchain-community==0.0.38
langchain-huggingface==0.0.3
langchain-text-splitters==0.0.1
faiss-cpu==1.8.0


unstructured==0.14.0
scikit-learn==1.4.0


numpy==1.26.0